Don't abort on errors when importing migration files
This is a quick fix to allow processing to continue on new migration files even
if there are old migrations that, for whatever reason, can no longer be imported

The better solution would be not to load the migration script code until it is
required. This would mean it would correctly abort when you try to apply the
migration, but not generate any errors if the migration has already been
applied. It would also improve startup time when there are large numbers
of migrations.